refactor(039-07): split global audio provider 982→79 lines into three hooks
extracted useAudioInit, useAudioTrackLoader, useAudioPlayback from
GlobalAudioProvider; provider reduced from 982 to 79 lines (target <200)
